Improved Key- and TimeSignature; Note 'diacritics'
The KeySignature object was updated: it now contains not only the cleff
variable, but also the actual key signature information as an integer
from -7 to 7 (the same way as in the MIDI specification).

The tokenizer for the TimeSignature object now also considers time
fractions with '12' as a numerator and/or '16' as a denominator.
The object per se hasn't been changed. A circle_of_fifths dictionary
has been added to semantics.coffee for eventual use in the MIDI
conversion or to determine the two possible keys ('major / minor').

The Note object now supports note diacritics, i.e. accidentals,
articulations, and the note dot. It represents them as a list of
strings. The get_name method still ignores this list, though. It still
has to be decided if the note accidentals are going to be displayed
or interpreted with respect to the actual key signature, so that the
key signature automatically affects all succeeding Note objects.