Dynamics, repetitions, octavation
Dynamics were added as a separate object from notes. Repetitions exist
with all available referential marks (with a corresponding *closing*
object for the repeat sections). Octavation of chords was also added as
a separate object (with a corresponding closing object). When
translating to MIDI, this should be checked for illegal
parenthesization.