Rests, ornamentation, pedal marks
A parent class 'TimeInterval' was added to the Note class and the new
Rest class. Rests also allow increments of duration via '~'.

Ornamentation was added to the note diacritics. The arpeggio is now
applicable to chords.

Pedal marks are represented as different objects. These also have to be
checked for valid parenthesization in the MIDI conversion.

There is also some minor renaming and restructuring.