Key signature + accidentals behavior
If an accidental is added and it was already in the key signature, it is
not a double accidental, but a simple accidental. If the opposite
accidental is added, it is not a natural, but the new accidental itself.

A small bug in the get_name method of Note was fixed.