Removed length_base class variable
The method Note.set_pitch changed length_base when the note was a quarter
note. However, the variable is reassigned in the base class' TimeInterval
constructor, which turns every quarter note into an eighth note. The note
exponent should be the only parameter affecting the note length and the
distinction of quarter and eighth notes should happen a priori (in the parser
function itself).