Separated translator and added file writer
It was not a good idea to concatenate the translator together with the parser:
it was polluting the namespace.

A file writer "write.py" was added, which can be invoked separately or via the
file watcher "observe.py". It demands the existence of the directory
"desktop/io" and a "in.txt" file inside it. The parser output file, the
translator/music21-wrapper output file, a midi file and a musicXML file are
written in the same directory.

Appart from that, the structure of the translator and the object definition in
the parser was completed, though not yet thoroughly tested.